reorganise branch window
Make the panning more obvious by moving the buttons outside of the top
pane and having a inset frame filling the whole bottom pane.

Make the bottom pane scrollable as a whole, so more space is available
to see long commit messages and so pathological revisions with many
parents do not prevent seeing the commit message.