Make "vizualize" use the GUI progress bar defined in the parent 'ui' module.
Update the progress bar so that one need not specify both the total steps and 
the current step with every update.

Remove extraneous signaling and put related progress-bar element close to its
use.

TODO:  Given a nested-progress-bar object, one should be able to use it as a
factory of sub-bars.